+++ CHANGELOG.md
@@ -3,18 +3,24 @@
 
 ## [Unreleased]
 ### Added
+### Fixed
+### Changed
+### Deprecated
+### Removed
+### Security
+
+
+## [1.1.0]
+### Added
 - add py3.7 to tox and travis tests
 ### Fixed
 - fixed sync issues with django-peeringdb and django3 (peeringdb/django-peeringdb#37)
 - fixed sync issues with django-peeringdb and mysql (#41)
 - better data for tests (#40)
 
-### Changed
-### Deprecated
 ### Removed
 - remove py2.7 support (#37)
 - remove py3.4 tests (py3.4 EOL reached)
-### Security
 
 
 ## [1.0.0] 2019-10-29